refactor prop into multiple files
